remove old homogenization algorithms
